revive `BooleanParser` and remove `TypeConvertParser`
